Split Condition/CondTree out of GenericPackageDescription.
